research_agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.28
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-core==0.3.0
tavily-python==0.5.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
structlog==24.4.0
httpx==0.27.2
sse-starlette==2.1.3
